[Keyless] Add "accounts" directory to the pepper service.
