+++ keyless/pepper/service/src/lib.rs
@@ -1,8 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod account_db;
-pub mod account_managers;
+pub mod accounts;
 pub mod dedicated_handlers;
 pub mod error;
 pub mod external_resources;
